Refactor query parameter
