+++ openapi-core/src/main/java/io/ballerina/openapi/core/generators/client/FunctionSignatureGeneratorImp.java
@@ -37,10 +37,10 @@ import io.ballerina.compiler.syntax.tree.SyntaxKind;
 import io.ballerina.compiler.syntax.tree.Token;
 import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
 import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
-import io.ballerina.openapi.core.GeneratorUtils;
-import io.ballerina.openapi.core.exception.BallerinaOpenApiException;
+import io.ballerina.openapi.core.generators.common.GeneratorUtils;
+import io.ballerina.openapi.core.generators.common.exception.BallerinaOpenApiException;
 import io.ballerina.openapi.core.generators.document.DocCommentsGenerator;
-import io.ballerina.openapi.core.generators.schema.BallerinaTypesGenerator;
+import io.ballerina.openapi.core.generators.type.BallerinaTypesGenerator;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.headers.Header;
@@ -80,30 +80,29 @@ import static io.ballerina.compiler.syntax.tree.SyntaxKind.COMMA_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.EQUAL_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.OPEN_PAREN_TOKEN;
 import static io.ballerina.compiler.syntax.tree.SyntaxKind.RETURNS_KEYWORD;
-import static io.ballerina.openapi.core.ErrorMessages.invalidPathParamType;
-import static io.ballerina.openapi.core.GeneratorConstants.APPLICATION_OCTET_STREAM;
-import static io.ballerina.openapi.core.GeneratorConstants.ARRAY;
-import static io.ballerina.openapi.core.GeneratorConstants.BOOLEAN;
-import static io.ballerina.openapi.core.GeneratorConstants.EMPTY_RECORD;
-import static io.ballerina.openapi.core.GeneratorConstants.HTTP_REQUEST;
-import static io.ballerina.openapi.core.GeneratorConstants.INTEGER;
-import static io.ballerina.openapi.core.GeneratorConstants.NILLABLE;
-import static io.ballerina.openapi.core.GeneratorConstants.NUMBER;
-import static io.ballerina.openapi.core.GeneratorConstants.OBJECT;
-import static io.ballerina.openapi.core.GeneratorConstants.PAYLOAD;
-import static io.ballerina.openapi.core.GeneratorConstants.REQUEST;
-import static io.ballerina.openapi.core.GeneratorConstants.SQUARE_BRACKETS;
-import static io.ballerina.openapi.core.GeneratorConstants.STRING;
-import static io.ballerina.openapi.core.GeneratorConstants.X_BALLERINA_DEPRECATED_REASON;
-import static io.ballerina.openapi.core.GeneratorUtils.convertOpenAPITypeToBallerina;
-import static io.ballerina.openapi.core.GeneratorUtils.escapeIdentifier;
-import static io.ballerina.openapi.core.GeneratorUtils.extractReferenceType;
-import static io.ballerina.openapi.core.GeneratorUtils.getBallerinaMediaType;
-import static io.ballerina.openapi.core.GeneratorUtils.getOpenAPIType;
-import static io.ballerina.openapi.core.GeneratorUtils.getValidName;
-import static io.ballerina.openapi.core.GeneratorUtils.isArraySchema;
-import static io.ballerina.openapi.core.GeneratorUtils.isComposedSchema;
-import static io.ballerina.openapi.core.GeneratorUtils.isObjectSchema;
+import static io.ballerina.openapi.core.generators.common.ErrorMessages.invalidPathParamType;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.ARRAY;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.BOOLEAN;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.EMPTY_RECORD;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.HTTP_REQUEST;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.INTEGER;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.NILLABLE;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.NUMBER;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.OBJECT;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.PAYLOAD;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.REQUEST;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.SQUARE_BRACKETS;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.STRING;
+import static io.ballerina.openapi.core.generators.common.GeneratorConstants.X_BALLERINA_DEPRECATED_REASON;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.convertOpenAPITypeToBallerina;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.escapeIdentifier;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.extractReferenceType;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.getBallerinaMediaType;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.getOpenAPIType;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.getValidName;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.isArraySchema;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.isComposedSchema;
+import static io.ballerina.openapi.core.generators.common.GeneratorUtils.isObjectSchema;
 
 /**
  * This util class uses for generating {@link FunctionSignatureNode} for given OAS
